generative_ai/rag/chatbot_pdf/requirements.txt
streamlit
langchain
openai
faiss-cpu
pypdf2
python-dotenv
huggingface-hub
